chore: clean up dead code and legacy artifacts
- Remove outdated analysis docs (FLOW_ARCHITECTURE_ANALYSIS.md,
  FLOW_REFACTORING_DESIGNS.md, FLOW_REFACTORING_CLEAN.md) that
  reference files no longer in codebase
- Inline types.ts into AgentRunFlowRunner.ts - types.ts was a shallow
  module with only 2 type definitions
- Update FLOW_REFACTORING_STATUS.md to remove reference to deleted doc
- Reduce common/ from 5 files to 4 files
