+++ src/agent/implementations/flows/common/index.ts
@@ -6,8 +6,8 @@
  * - AgentRunShared: Generic shared state container for all flows
  * - AgentRunFlowOptions: Options for runAgentFlow
  * - runAgentFlow: Flow execution with hooks
- * - StandardFinalizeNode: Standard finalize node class
  * - InitExecResult, NodeExecResult: Shared result types for nodes
+ * - StandardFinalizeNode, FinalizeContext: Finalize node class and context type
  *
  * Note: Schemas are defined in respective run flow files.
  * Note: Init nodes are inlined in each flow - no factory needed.
@@ -16,4 +16,3 @@
 export * from './AgentLifecycle';
 export * from './AgentRunFlowRunner';
 export * from './createFinalizeNode';
-export * from './types';
